Moved bookreservation_details to the Pages folder
