+++ src/app_feup/lib/view/Pages/book_reservation_details.dart
